pyproject.toml
[project]
name = "crewai-mcp-dnd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "crewai[tools]>=0.118.0",
    "fastmcp>=2.8.0",
    "glom>=24.11.0",
    "langchain>=0.3.25",
    "langfuse>=3.0.1",
    "openai>=1.86.0",
    "openlit>=1.33.23",
    "opentelemetry-exporter-otlp>=1.34.1",
    "opentelemetry-sdk>=1.34.1",
]
